chore(library/defeq_simplifier): move to new module inside library/tactic
