+++ src/CMakeLists.txt
@@ -331,6 +331,8 @@ add_subdirectory(library)
 set(LEAN_OBJS ${LEAN_OBJS} $<TARGET_OBJECTS:library>)
 add_subdirectory(library/tactic)
 set(LEAN_OBJS ${LEAN_OBJS} $<TARGET_OBJECTS:tactic>)
+add_subdirectory(library/tactic/defeq_simplifier)
+set(LEAN_OBJS ${LEAN_OBJS} $<TARGET_OBJECTS:defeq_simplifier>)
 add_subdirectory(library/definitional)
 set(LEAN_OBJS ${LEAN_OBJS} $<TARGET_OBJECTS:definitional>)
 # add_subdirectory(library/blast)
