+++ src/frontends/lean/builtin_cmds.cpp
@@ -29,8 +29,8 @@ Author: Leonardo de Moura
 #include "library/type_context.h"
 #include "library/legacy_type_context.h"
 #include "library/reducible.h"
-#include "library/defeq_simp_lemmas.h"
-#include "library/defeq_simplifier.h"
+#include "library/tactic/defeq_simplifier/defeq_simp_lemmas.h"
+#include "library/tactic/defeq_simplifier/defeq_simplifier.h"
 #include "library/vm/vm.h"
 #include "library/vm/vm_string.h"
 #include "library/compiler/vm_compiler.h"
