+++ src/library/blast/blast.cpp
@@ -15,8 +15,8 @@ Author: Leonardo de Moura
 #include "kernel/type_checker.h"
 #include "library/replace_visitor.h"
 #include "library/util.h"
-#include "library/defeq_simp_lemmas.h"
-#include "library/defeq_simplifier.h"
+#include "library/tactic/defeq_simplifier/defeq_simp_lemmas.h"
+#include "library/tactic/defeq_simplifier/defeq_simplifier.h"
 #include "library/trace.h"
 #include "library/reducible.h"
 #include "library/class.h"
