+++ src/library/tactic/defeq_simplifier/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(defeq_simplifier OBJECT init_module.cpp defeq_simp_lemmas.cpp defeq_simplifier.cpp)
