+++ src/library/tactic/defeq_simplifier/init_module.h
@@ -1,0 +1,11 @@
+/*
+Copyright (c) 2016 Daniel Selsam. All rights reserved.
+Released under Apache 2.0 license as described in the file LICENSE.
+Author: Daniel Selsam
+*/
+#pragma once
+
+namespace lean {
+void initialize_defeq_simplifier_module();
+void finalize_defeq_simplifier_module();
+}
